+++ features/settings/impl/src/commonMain/kotlin/app/campfire/settings/SleepSettingsImpl.kt
@@ -97,6 +97,10 @@ class SleepSettingsImpl(
   private val autoRewindAmountProperty = durationSetting(KEY_AUTO_REWIND_AMOUNT, DefaultAutoRewindAmount)
   override var autoRewindAmount: Duration by autoRewindAmountProperty
   override fun observeAutoRewindAmount(): StateFlow<Duration> = autoRewindAmountProperty.observe()
+
+  private val fadeOutDurationProperty = durationSetting(KEY_FADE_OUT_DURATION, SleepSettings.DefaultFadeOutDuration)
+  override var fadeOutDuration: Duration by fadeOutDurationProperty
+  override fun observeFadeOutDuration(): StateFlow<Duration> = fadeOutDurationProperty.observe()
 }
 
 private const val KEY_LAST_SET_SLEEP_TIMER = "pref_last_set_sleep_timer"
@@ -108,6 +112,7 @@ private const val KEY_AUTO_SLEEP_END = "pref_sleep_auto_timer_end"
 private const val KEY_AUTO_SLEEP_TIMER = "pref_auto_sleep_timer"
 private const val KEY_AUTO_REWIND_ENABLED = "pref_auto_rewind_enabled"
 private const val KEY_AUTO_REWIND_AMOUNT = "pref_auto_rewind_amount"
+private const val KEY_FADE_OUT_DURATION = "pref_sleep_fade_out_duration"
 
 private const val DefaultShakeToResetEnabled = false
 private const val DefaultAutoSleepTimerEnabled = false